Added a restore option to Mod Loader for PC so now any PC player with something like rando can just click restore whether patching or using mod loader to "uninstall" the neccesarry rando stuff to go back to vanilla. Also added multigame functionaility to mod manager. Still need to make a setup wizard window for extracting all PC games in the kh 1.5 2.5 collection but bools are created in mods-manager.yml that can be set to true to make the extract button do more than just kh2. Also progress ar doesnt seem to work correctly anymore.

+++ OpenKh.Tools.ModsManager/ViewModels/MainViewModel.cs
@@ -41,8 +41,17 @@ namespace OpenKh.Tools.ModsManager.ViewModels
         private bool _pc;
         private bool _panaceaInstalled;
         private bool _devView;
-        private string _quickLaunch = "kh2";
+        private string _launchGame = "kh2";
         private int _wizardVersionNumber = 1;
+        private string[] executable = new string[]
+        {
+            "KINGDOM HEARTS II FINAL MIX.exe",
+            "KINGDOM HEARTS FINAL MIX.exe",
+            "KINGDOM HEARTS Birth by Sleep FINAL MIX.exe",
+            "KINGDOM HEARTS Re_Chain of Memories.exe",
+            ""
+        };
+        private int launchExecutable = 0;
 
         private const string RAW_FILES_FOLDER_NAME = "raw";
         private const string ORIGINAL_FILES_FOLDER_NAME = "original";
@@ -91,7 +100,7 @@ namespace OpenKh.Tools.ModsManager.ViewModels
         public Visibility PatchVisible => PC && !PanaceaInstalled || PC && DevView ? Visibility.Visible : Visibility.Collapsed;
         public Visibility ModLoader => !PC || PanaceaInstalled ? Visibility.Visible : Visibility.Collapsed;
         public Visibility HideStop => !PC ? Visibility.Visible : Visibility.Collapsed;
-        public Visibility HideQuickLaunch => PC && PanaceaInstalled && DevView ? Visibility.Visible : Visibility.Collapsed;
+        public Visibility isPC => PC ? Visibility.Visible : Visibility.Collapsed;
 
         public bool DevView
         {
@@ -101,7 +110,6 @@ namespace OpenKh.Tools.ModsManager.ViewModels
                 _devView = value;
                 ConfigurationService.DevView = DevView;
                 OnPropertyChanged(nameof(PatchVisible));
-                OnPropertyChanged(nameof(HideQuickLaunch));
             }
         }
         public bool PanaceaInstalled
@@ -125,14 +133,15 @@ namespace OpenKh.Tools.ModsManager.ViewModels
                 OnPropertyChanged(nameof(ModLoader));
                 OnPropertyChanged(nameof(PatchVisible));
                 OnPropertyChanged(nameof(HideStop));
+                OnPropertyChanged(nameof(isPC));
             }
         }  
 
-        public int QuickLaunch
+        public int GametoLaunch
         {
             get
             {
-                switch (_quickLaunch)
+                switch (_launchGame)
                 {
                     case "kh2":
                         return 0;
@@ -150,22 +159,28 @@ namespace OpenKh.Tools.ModsManager.ViewModels
             }
             set
             {
+                launchExecutable = value;
                 switch (value)
                 {
                     case 0:
-                        _quickLaunch = "kh2";
+                        _launchGame = "kh2";
+                        ConfigurationService.LaunchGame = "kh2";
                         break;
                     case 1:
-                        _quickLaunch = "kh1";
+                        _launchGame = "kh1";
+                        ConfigurationService.LaunchGame = "kh1";
                         break;
                     case 2:
-                        _quickLaunch = "bbs";
+                        _launchGame = "bbs";
+                        ConfigurationService.LaunchGame = "bbs";
                         break;
                     case 3:
-                        _quickLaunch = "recom";
+                        _launchGame = "recom";
+                        ConfigurationService.LaunchGame = "recom";
                         break;
                     default:
-                        _quickLaunch = "off";
+                        _launchGame = "off";
+                        ConfigurationService.LaunchGame = "off";
                         break;
                 }
             }
@@ -189,6 +204,7 @@ namespace OpenKh.Tools.ModsManager.ViewModels
 
         public MainViewModel()
         {
+            ConfigurationService.LaunchGame = _launchGame;
             if (ConfigurationService.GameEdition == 2)
             {
                 PC = true;
@@ -308,10 +324,10 @@ namespace OpenKh.Tools.ModsManager.ViewModels
                 await RunGame();
             });
 
-            RestoreCommand = new RelayCommand(async _ =>
+            RestoreCommand = new RelayCommand(async (patched) =>
             {
                 ResetLogWindow();
-                await RestoreGame();
+                await RestoreGame(Convert.ToBoolean(patched));
                 CloseAllWindows();
             });
             BuildAndRunCommand = new RelayCommand(async _ =>
@@ -464,7 +480,7 @@ namespace OpenKh.Tools.ModsManager.ViewModels
                     {
                         if (ConfigurationService.PanaceaInstalled)
                         {
-                            File.AppendAllText(Path.Combine(ConfigurationService.PcReleaseLocation, "panacea_settings.txt"), "\nquick_launch=" + _quickLaunch);
+                            File.AppendAllText(Path.Combine(ConfigurationService.PcReleaseLocation, "panacea_settings.txt"), "\nquick_launch=" + _launchGame);
                         }                        
                         processStartInfo = new ProcessStartInfo
                         {
@@ -476,14 +492,14 @@ namespace OpenKh.Tools.ModsManager.ViewModels
                     {
                         processStartInfo = new ProcessStartInfo
                         {
-                            FileName =  Path.Combine(ConfigurationService.PcReleaseLocation, "KINGDOM HEARTS II FINAL MIX.exe"),
+                            FileName =  Path.Combine(ConfigurationService.PcReleaseLocation, executable[launchExecutable]),
                             WorkingDirectory = ConfigurationService.PcReleaseLocation,
                             UseShellExecute = false,
                         };
                         if (processStartInfo == null || !File.Exists(processStartInfo.FileName))
                         {
                             MessageBox.Show(
-                                "Unable to start game. Please make sure youre Kingdom Hearts executable is correctly named and in the correct folder.",
+                                "Unable to start game. Please make sure your Kingdom Hearts executable is correctly named and in the correct folder.",
                                 "Run error", MessageBoxButton.OK, MessageBoxImage.Error);
                             CloseAllWindows();
                             return Task.CompletedTask;
@@ -587,25 +603,25 @@ namespace OpenKh.Tools.ModsManager.ViewModels
                 if (ConfigurationService.GameEdition == 2)
                 {
                     // Use the package map file to rearrange the files in the structure needed by the patcher
-                    var packageMapLocation = Path.Combine(ConfigurationService.GameModPath, "patch-package-map.txt");
+                    var packageMapLocation = Path.Combine(ConfigurationService.GameModPath, _launchGame , "patch-package-map.txt");
                     var packageMap = File
                         .ReadLines(packageMapLocation)
                         .Select(line => line.Split(" $$$$ "))
                         .ToDictionary(array => array[0], array => array[1]);
 
-                    var patchStagingDir = Path.Combine(ConfigurationService.GameModPath, "patch-staging");
+                    var patchStagingDir = Path.Combine(ConfigurationService.GameModPath, _launchGame, "patch-staging");
                     if (Directory.Exists(patchStagingDir))
                         Directory.Delete(patchStagingDir, true);
                     Directory.CreateDirectory(patchStagingDir);
                     foreach (var entry in packageMap)
                     {
-                        var sourceFile = Path.Combine(ConfigurationService.GameModPath, entry.Key);
+                        var sourceFile = Path.Combine(ConfigurationService.GameModPath, _launchGame, entry.Key);
                         var destFile = Path.Combine(patchStagingDir, entry.Value);
                         Directory.CreateDirectory(Path.GetDirectoryName(destFile));
                         File.Move(sourceFile, destFile);
                     }
 
-                    foreach (var directory in Directory.GetDirectories(ConfigurationService.GameModPath))
+                    foreach (var directory in Directory.GetDirectories(Path.Combine(ConfigurationService.GameModPath, _launchGame)))
                         if (!"patch-staging".Equals(Path.GetFileName(directory)))
                             Directory.Delete(directory, true);
 
@@ -736,31 +752,45 @@ namespace OpenKh.Tools.ModsManager.ViewModels
             });
         }
 
-        private async Task RestoreGame()
+        private async Task RestoreGame(bool patched)
         {
             await Task.Run(() =>
             {
                 if (ConfigurationService.GameEdition == 2)
-                {
-                    if (!Directory.Exists(Path.Combine(ConfigurationService.PcReleaseLocation, "BackupImage")))
+                {                        
+                    if(patched)
                     {
-                        Log.Warn("backup folder cannot be found! Cannot restore the game.");
-                    }
-
-                    else
-                    {
-                        foreach (var file in Directory.GetFiles(Path.Combine(ConfigurationService.PcReleaseLocation, "BackupImage")).Where(x => x.Contains(".pkg")))
+                        if (!Directory.Exists(Path.Combine(ConfigurationService.PcReleaseLocation, "BackupImage")))
                         {
-                            Log.Info($"Restoring Package File {file.Replace(".pkg", "")}");
+                            Log.Warn("backup folder cannot be found! Cannot restore the game.");
+                        }
+                        else
+                        {
+                            foreach (var file in Directory.GetFiles(Path.Combine(ConfigurationService.PcReleaseLocation, "BackupImage")).Where(x => x.Contains(".pkg")))
+                            {
+                                Log.Info($"Restoring Package File {file.Replace(".pkg", "")}");
 
-                            var _fileBare = Path.GetFileName(file);
-                            var _trueName = Path.Combine(ConfigurationService.PcReleaseLocation, "Image", ConfigurationService.PcReleaseLanguage, _fileBare);
+                                var _fileBare = Path.GetFileName(file);
+                                var _trueName = Path.Combine(ConfigurationService.PcReleaseLocation, "Image", ConfigurationService.PcReleaseLanguage, _fileBare);
 
-                            File.Delete(Path.ChangeExtension(_trueName, "hed"));
-                            File.Delete(_trueName);
+                                File.Delete(Path.ChangeExtension(_trueName, "hed"));
+                                File.Delete(_trueName);
 
-                            File.Copy(file, _trueName);
-                            File.Copy(Path.ChangeExtension(file, "hed"), Path.ChangeExtension(_trueName, "hed"));
+                                File.Copy(file, _trueName);
+                                File.Copy(Path.ChangeExtension(file, "hed"), Path.ChangeExtension(_trueName, "hed"));
+                            }
+                        }
+
+                    }
+                    if (Directory.Exists(ConfigurationService.GameModPath))
+                    {
+                        try
+                        {
+                            Directory.Delete(Path.Combine(ConfigurationService.GameModPath, _launchGame), true);
+                        }
+                        catch (Exception ex)
+                        {
+                            Log.Warn("Unable to fully clean the mod directory:\n{0}", ex.Message);
                         }
                     }
                 }
